+++ History.md
@@ -2,6 +2,7 @@ Unpublished
 ===========
 
   * callback for writeI2cBlock now gets 3 arguments (err, bytesWritten, buffer)
+  * added writeQuick and writeQuickSync
 
 0.8.0 / Dec 19 2014
 ===================